requirements.txt
asn1crypto==0.23.0
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage==0.36.0
Babel==2.5.1
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.1
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Table==0.4.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pip==9.0.1
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
setuptools==36.6.0
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
wheel==0.30.0
WTForms==2.1
